Move input.h from components/ to systems/ (refs #1194)
InputState/TouchSlot/InputSource are hardware-capture state owned by
input_system as a registry ctx singleton — they are never emplaced on
entities, they are not queried via views, and only quit_requested is
read outside the input subsystem. That makes them per-system state, not
component data, exactly the pattern the #1194 audit asked us to relocate
beside the owning system.

Changes:

- New app/systems/input.h with the same declarations (one breadcrumb
  comment added pointing at #1194 as the move's rationale).
- Delete app/components/input.h.
- Update every include site: app/game_loop.cpp, app/systems/{input,
  game_state_system,game_state_routing}.cpp, the 5 ui/screen_controllers
  that read quit_requested or the InputState singleton, plus
  tests/test_helpers.h, tests/test_lifecycle.cpp, and
  benchmarks/bench_systems.cpp.
- Refresh the #1196 breadcrumb in input_system_private.h: keep the
  historical "extracted out of InputState (#1196)" attribution and
  note that InputState itself later followed (this PR, #1194).
- design-docs/architecture.md: move InputState/TouchSlot/InputSource
  snippet header from components/input.h to systems/input.h; refresh
  the directory tree (input.h now lives under systems/).
- design-docs/rguilayout-portable-c-integration.md: update the
  example title_screen_controller include.

Validation:

- Unity Clang `-Wall -Wextra -Werror`: clean, 0 warnings.
- Non-unity Clang `-Wall -Wextra -Werror`: clean, 0 warnings.
- ./build/shapeshifter_tests: 901 test cases / 2957 assertions pass.
- ./build-nounity/shapeshifter_tests: 901 test cases / 2957 assertions pass.

Refs #1194.
