+++ app/ui/screen_controllers/gameplay_hud_screen_controller.cpp
@@ -2,7 +2,7 @@
 
 #include "../../components/game_state.h"
 #include "../../components/energy_bar.h"
-#include "../../components/input.h"
+#include "../../systems/input.h"
 #include "../../systems/input_events.h"
 #include "../../components/obstacle.h"
 #include "../../components/player.h"
