+++ app/ui/screen_controllers/title_screen_controller.cpp
@@ -1,7 +1,7 @@
 // Title screen controller - bridges generated rguilayout output to game systems.
 
 #include "../../components/game_state.h"
-#include "../../components/input.h"
+#include "../../systems/input.h"
 #include "screen_controller_base.h"
 #include <raygui.h>
 #include "title_screen_dead_zones.h"
